GetCertificateCN(): if the CN is missing, fall back to the DNS SAN
if exactly one is given and use it as CN.

Currently Icinga 2 populates both the DNS SAN and the subject CN of a newly
issued certificate with the endpoint name. But the CN is limited to 64
characters. So endpoint names are. The only clean solution is omitting the
CN for too long names. Icinga 2 must be able to extract the endpoint name
from such certificates in the first place and exactly one DNS SAN is a legit
source.